add: determine ECP electrons from ORCA json

+++ pysisyphus/elem_data.py
@@ -3561,7 +3561,7 @@ INV_ATOMIC_NUMBERS = {num: elem for elem, num in ATOMIC_NUMBERS.items()}
 
 
 def nuc_charges_for_atoms(atoms):
-    return np.array([ATOMIC_NUMBERS[atom.lower()] for atom in atoms])
+    return np.array([ATOMIC_NUMBERS[atom.lower()] for atom in atoms], dtype=int)
 
 
 def atomic_numbers_for_atoms(atoms):
